moved globalSettings to src
