+++ src/mappers/graphStatsMapper.py
@@ -1,6 +1,7 @@
 import pandas as pd
 
 from src.analysis.methods.graphStats import GraphStats
+from src.globalSettings import GlobalSettings
 
 class GraphStatsMapper:
     @staticmethod
@@ -29,23 +30,10 @@ class GraphStatsMapper:
                 stats.componentSizeDistribution
                ]
 
-    @staticmethod
-    def colnames():
-        return [
-                "isolatedVerticeRatio",
-                "edgeDensity",
-                "density",
-                "meanEccentricity",
-                "giantComponent",
-                "meanDegree",
-                "componentCount",
-                "meanComponentSize"
-               ]
-
     @classmethod
     def toDataframe(cls, statList: list[GraphStats]) -> pd.DataFrame:
         data = [ cls.toList(stat) for stat in statList ]
-        df = pd.DataFrame(data=data, columns=pd.Index(cls.colnames()))
+        df = pd.DataFrame(data=data, columns=pd.Index(GlobalSettings().graphStatVectorNames))
         return df
