+++ src/models/dataset.py
@@ -10,7 +10,7 @@ from typing import List
 
 from .base import Base
 
-from src.creation.globalSettings import globalSettings
+from src.globalSettings import GlobalSettings
 
 class Dataset(Base):
     __tablename__ = "dataset"
@@ -20,8 +20,8 @@ class Dataset(Base):
                                                   default=uuid.uuid4)
     name: Mapped[str] =  mapped_column(String(30), nullable = False)
     description: Mapped[str] =  mapped_column(String(120), nullable = False)
-    version: Mapped[str] =  mapped_column(String(10), nullable = False, default=globalSettings().version)
-    username: Mapped[str] =  mapped_column(String(30), nullable = False, default=globalSettings().defaultUsername)
+    version: Mapped[str] =  mapped_column(String(10), nullable = False, default=GlobalSettings().version)
+    username: Mapped[str] =  mapped_column(String(30), nullable = False, default=GlobalSettings().defaultUsername)
     modificationDate: Mapped[date] =  mapped_column(Date, nullable = False, default=datetime.now())
     creationDate: Mapped[date] =  mapped_column(Date, nullable = False, default=datetime.now())
